requirements.txt
flask
requests
python-dotenv
psycopg2-binary

